Tweet A Complaint
The project aims at developing an instant messenger application for
windows desktops. It allows UNCC students to have the ability to
communicate with one another in an easy manner. Because of
unavailability of an instant messaging application, we are left with
only one option that is, dropping a formal email. Unfortunately, there
is one problem with having an email conversation. At times, you might
not get an immediate response to your email. To solve this problem, we
developed a messaging application that allows students to communicate
informally.